Revert "Enable nullability in most of the remaining files in BizHawk.Common"
This reverts commit 51b62629c61c0620db4c6b9f9f3db17115c361c4.

Bizhawk does not change much; the application's features are largely the same as they were years ago, with some core work being the only notable improvements from release to release.  Yet there is always a massive amount of churn in the codebase, with multiple contributors engaing in zealous refactoring costantly.  This has to stop.  It's too much effort to police and fix regressions, for no gain -- if the code was simply left as it was, it would still work the way it did yesterday, and that would be fine.

From now on, whenever I see a commit that:
1. contains a problem
2. does not provide any real bugfix or real feature for end users
I'm just going to revert it with no further comment.  It's not worth it to sort out the good from the bad, because of #2 -- there's no value at the end of the tunnel.

If the abusive refactorings continue, I'll start reverting all refactoring commits.

+++ src/BizHawk.Common/TempFileManager.cs
@@ -1,3 +1,5 @@
+#nullable disable
+
 using System;
 using System.Linq;
 using System.Collections.Generic;
@@ -15,7 +17,7 @@ namespace BizHawk.Common
 	{
 		// TODO - manage paths other than %temp%, make not static, or allow adding multiple paths to static instance
 
-		public static string GetTempFilename(string friendlyName, string? dotAndExtension = null, bool delete = true)
+		public static string GetTempFilename(string friendlyName, string dotAndExtension = null, bool delete = true)
 		{
 			string guidPart = Guid.NewGuid().ToString();
 			var fname = $"biz-{System.Diagnostics.Process.GetCurrentProcess().Id}-{friendlyName}-{guidPart}{dotAndExtension ?? ""}";
@@ -31,7 +33,7 @@ namespace BizHawk.Common
 		public static string RenameTempFilenameForDelete(string path)
 		{
 			string filename = Path.GetFileName(path);
-			var dir = Path.GetDirectoryName(path) ?? throw new NullReferenceException();
+			string dir = Path.GetDirectoryName(path);
 			if (!filename.StartsWith("biz-"))
 			{
 				throw new InvalidOperationException();
@@ -75,35 +77,37 @@ namespace BizHawk.Common
 
 				foreach(var di in dis)
 				{
-					FileInfo[] fis;
+					FileInfo[] fis = null;
 					try
 					{
 						fis = di.GetFiles("bizdelete-*");
 					}
 					catch
 					{
-						continue;
 					}
 
-					foreach (var fi in fis)
+					if (fis != null)
 					{
-						try
+						foreach (var fi in fis)
 						{
-							if (OSTailoredCode.IsUnixHost)
+							try
 							{
-								fi.Delete(); // naive deletion, Mono doesn't care
+								if (OSTailoredCode.IsUnixHost)
+								{
+									fi.Delete(); // naive deletion, Mono doesn't care
+								}
+								else
+								{
+									Win32Imports.DeleteFileW(fi.FullName); // SHUT. UP. THE. EXCEPTIONS.
+								}
 							}
-							else
+							catch
 							{
-								Win32Imports.DeleteFileW(fi.FullName); // SHUT. UP. THE. EXCEPTIONS.
 							}
-						}
-						catch
-						{
-						}
 
-						// try not to do more than one thing per frame
-						Thread.Sleep(100);
+							// try not to do more than one thing per frame
+							Thread.Sleep(100);
+						}
 					}
 				}
 
@@ -116,7 +120,7 @@ namespace BizHawk.Common
 		{
 		}
 
-		private static Thread? thread;
+		private static Thread thread;
 
 		public static void HelperSetTempPath(string path)
 		{
